move more traits to `rust_ir`
